refactor: delete and replace old pagination component
